backend/requirements.txt
gradium
sphn
# fastapi[all]
sounddevice 
numpy 
python-dotenv
openai
google-generativeai
mistralai
openai